updated dependencies and bookings refactor
